+++ alobe/event.h
@@ -13,15 +13,15 @@ class Actor;
 class Event
 {
     public:
-        Event(unsigned int step = 0);
-        void action(Actor & actor, unsigned int step); // Trigger event action on an actor with a given step
+        Event(unsigned long step = 0);
+        void action(Actor & actor, unsigned long step); // Trigger event action on an actor with a given step
 
     protected:
         virtual void filteredAction(Actor & actor) = 0; // Filtered action is called only if step == triggerStep
                                                         // This method is protected so that it can be called
                                                         // from abstract inherited classes
     protected:
-        unsigned int my_triggerStep;                    // This property is protected so that it can be called
+        unsigned long my_triggerStep;                   // This property is protected so that it can be called
                                                         // from abstract inherited classes
 };
