+++ alobe/simulation.h
@@ -19,16 +19,16 @@ class Simulation
 {
     public:
         Simulation(unique_ptr<Stepper> stepper, unique_ptr<Land> land);
-        void addBeing(shared_ptr<Being> being, unsigned int x, unsigned int y); // Add a being
-        void randomBeing(string name = "");                                     // Generate a random being
-        void toStep(unsigned int step);                                         // Calculate simulation steps until reaching 'step'
-        void nextStep();                                                        // Calculate the simulation next step
-        Stepper * getStepper() const;                                           // Getter for the stepper
-        Land * getLand() const;                                                 // Getter for the land
-        vector<shared_ptr<Being> > getBeings() const;                           // Getter for the beings
+        void addBeing(shared_ptr<Being> being, unsigned long x, unsigned long y); // Add a being
+        void randomBeing(string name = "");                                       // Generate a random being
+        void toStep(unsigned long step);                                          // Calculate simulation steps until reaching 'step'
+        void nextStep();                                                          // Calculate the simulation next step
+        Stepper * getStepper() const;                                             // Getter for the stepper
+        Land * getLand() const;                                                   // Getter for the land
+        vector<shared_ptr<Being> > getBeings() const;                             // Getter for the beings
 
     private:
-        void nextStepCallback();                                                // Called at the end of the nextStep method
+        void nextStepCallback();                                                  // Called at the end of the nextStep method
 
     private:
         unique_ptr<Stepper> my_stepper;
